Compilation support for Windows (MSVC, Clang and GCC)

+++ examples/windowed.c
@@ -55,7 +55,7 @@ int main(void) {
     // Align the sequences!
     printf("Aligning '%s' and '%s' using Hirschberg\n", pattern, text);
 
-    status = quicked_align(&aligner, pattern, strlen(pattern), text, strlen(text));
+    status = quicked_align(&aligner, pattern, (int)strlen(pattern), text, (int)strlen(text));
     if (quicked_check_error(status)) {
         fprintf(stderr, "%s", quicked_status_msg(status));
         return 1;
