+++ quicked_utils/src/vector.c
@@ -65,7 +65,7 @@ void vector_reserve(
     const uint64_t num_elements,
     const bool zero_mem) {
   if (vector->elements_allocated < num_elements) {
-    const uint64_t proposed = (float)vector->elements_allocated*VECTOR_EXPAND_FACTOR;
+    const uint64_t proposed = (uint64_t)((float)vector->elements_allocated*VECTOR_EXPAND_FACTOR);
     vector->elements_allocated = num_elements>proposed?num_elements:proposed;
     vector->memory = realloc(vector->memory,vector->elements_allocated*vector->element_size);
     if (!vector->memory) {
