+++ tools/align_benchmark/align_benchmark.c
@@ -81,9 +81,9 @@ bool align_benchmark_read_input(
   // Parameters
   int line1_length=0, line2_length=0;
   // Read queries
-  line1_length = getline(line1,line1_allocated,input_file);
+  line1_length = (int)getline(line1,line1_allocated,input_file);
   if (line1_length==-1) return false;
-  line2_length = getline(line2,line2_allocated,input_file);
+  line2_length = (int)getline(line2,line2_allocated,input_file);
   if (line1_length==-1) return false;
   // Configure input
   align_input->sequence_id = seqs_processed;
@@ -276,9 +276,9 @@ void align_benchmark_parallel(void) {
         sequence_offset_t* const offset = sequence_buffer->offsets + seq_idx;
         align_input[tid].sequence_id = seqs_processed;
         align_input[tid].pattern = sequence_buffer->buffer + offset->pattern_offset;
-        align_input[tid].pattern_length = offset->pattern_length;
+        align_input[tid].pattern_length = (int)(offset->pattern_length);
         align_input[tid].text = sequence_buffer->buffer + offset->text_offset;
-        align_input[tid].text_length = offset->text_length;
+        align_input[tid].text_length = (int)(offset->text_length);
         // Execute the selected algorithm
         align_benchmark_run_algorithm(align_input+tid);
       }
